requirements.txt
flask
flask-cors
langchain
python-dotenv
langchain-core
langchain-groq
langgraph
langgraph-checkpoint
langsmith
langgraph-prebuilt
langgraph-sdk
groq
python-dotenv
requests
transformers
mysql-connector-python
werkzeug
PyJWT